+++ src/Pages/NewProducts/Components/NewItemContainer/NewItemContainer.scss
@@ -1,0 +1,38 @@
+%flexCenter {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.NewItemContainer {
+  @extend %flexCenter;
+  width: 640px;
+  height: 100vh;
+  margin: 0px auto;
+  border-top: 10px solid rgb(235, 235, 235);
+
+  .ProductsList {
+    width: 570px;
+    height: 100vh;
+    margin-top: 100px;
+
+    .theme {
+      font-size: 15px;
+      font-weight: 400;
+      color: #9a9a9e;
+    }
+    h3 {
+      font-size: 28px;
+      font-weight: bold;
+      color: #000000;
+      margin-top: 10px;
+    }
+    .ProductsCardGrid {
+      display: grid;
+      grid-template-columns: repeat(2, 1fr);
+      gap: 20px;
+      width: 100%;
+      height: 100vh;
+    }
+  }
+}
