+++ src/Pages/NewProducts/Components/SliderContainer/SliderContainer.scss
@@ -1,0 +1,5 @@
+.SliderContainer {
+  width: 640px;
+  height: 360.09px;
+  margin: auto;
+}
